+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+cinecturlink2 2.2 - 2023.10.18
+===========================================================
+* Require Dotclear 2.28
+* Require PHP 8.1
+* Upgrade plugin activityReport
+
 cinecturlink2 2.1 - 2023.10.15
 ===========================================================
 * Require Dotclear 2.28